Move Last.fm logic into separate helper class
- Split AudioScrobblerSevice into two classes
- Change api to https in last.fm Audio Scrobbler
